requirements.txt
python-whois
python-nmap